login-app/backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0